service.js created, add tab id in localestorage...

+++ Django project/facturasieli/views/service/service_view.py
@@ -33,6 +33,9 @@ def display_service(request, company_id):
 
 # delete selected service
 def delete_service(request, service_id):
+    if not request.user.is_authenticated:
+        return HttpResponseRedirect(reverse('facturasieli:custom_log_in'))
+    
     service = get_object_or_404(Service, pk=service_id)
     service.delete()
